+++ pages/u_benef.php
@@ -24,7 +24,7 @@ include '../modules/u_nav_bar.php';
 
                 if (isset($_SESSION["user_id"])) {
 
-                    $my_benef = $bdd->prepare('SELECT b.benef_id, u.user_last_name, u.user_first_name, a.account_rib
+                    $my_benef = $bdd->prepare('SELECT b.benef_id, u.user_last_name, u.user_first_name, a.account_rib, b.benef_status
                     FROM users u
                     INNER JOIN accounts a ON a.account_user_id = u.user_id
                     INNER JOIN beneficiaries b ON b.account_id_2 = a.account_id
@@ -43,6 +43,7 @@ include '../modules/u_nav_bar.php';
                                     echo "<th>Nom</th>";
                                     echo "<th>Prénom</th>";
                                     echo "<th>RIB</th>";
+                                    echo "<th>Status</th>";
                                 echo "</tr>";
                             echo "</thead>";
                             echo "<tbody>";
@@ -51,10 +52,13 @@ include '../modules/u_nav_bar.php';
                                     echo "<td>" . $row['user_last_name'] . "</td>";
                                     echo "<td>" . $row['user_first_name'] . "</td>";
                                     echo "<td>" . $row['account_rib'] . "</td>";
+                                    echo "<td>" . $row['benef_status'] . "</td>";
                                     echo "<td>";
                                         echo "<div class='btn-group'>";
+                                        if ($row['benef_status'] == 'valide') {
                                         $row_account_id = $row['benef_id'];
                                         echo "<a href='../controllers/pdo_u_delete_acc.php?acc_id=".$row_account_id."'><button type='button' class='btn btn-primary'>Supprimer </button></a></div>";
+                                        }
                                     echo "</td>";
                                 echo "</tr>";
                             }
